Move spire-server api bits behind legacy build flag
