+++ cmd/spire-identity-exchange-server/main.go
@@ -5,19 +5,14 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
-	"net"
 	"os"
 	"os/signal"
 	"syscall"
 
-	githuboidc "github.com/spiffe/spire-identity-exchange/internal/github-oidc"
-	k8ssatoken "github.com/spiffe/spire-identity-exchange/internal/k8s-sa-token"
-	"github.com/spiffe/spire-identity-exchange/internal/cache"
 	"github.com/spiffe/spire-identity-exchange/internal/config"
 	prommetrics "github.com/spiffe/spire-identity-exchange/internal/metrics/prometheus"
 	"github.com/spiffe/spire-identity-exchange/internal/service"
 	"github.com/spiffe/spire-identity-exchange/pkg/validator"
-	"github.com/spiffe/spire/cmd/spire-server/util"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 )
@@ -65,23 +60,9 @@ func run() error {
 		logger = *rawLogger
 	}
 
-	// Create SPIRE client
-	socketPath := cfg.SPIRE.UnixSocketPath
-	var spireClient util.ServerClient
-	if cfg.Server.Port != 0 && (cfg.GitHubOIDC.Enabled || cfg.K8sSAToken.Enabled) {
-		if socketPath == "" {
-			logger.Error("unix_socket_path is required")
-			return fmt.Errorf("unix_socket_path is required")
-		}
-		spireClient, err = util.NewServerClient(&net.UnixAddr{
-			Name: socketPath,
-			Net:  "unix",
-		})
-		if err != nil {
-			logger.Error("failed to connect to SPIRE server via Unix socket", zap.Error(err))
-			return err
-		}
-		defer spireClient.Release()
+	spireClient, err := createSpireClient(cfg, &logger)
+	if err != nil {
+		return err
 	}
 
 	// Initialize metrics (includes process and Go runtime metrics)
@@ -102,53 +83,17 @@ func run() error {
 	}()
 	logger.Info("Metrics server initialized with runtime metrics", zap.Int("port", cfg.Server.MetricsPort))
 
-	var githubOIDCValidator validator.TokenValidator
-	var k8sSATokenValidator validator.TokenValidator
+	githubOIDCValidator, k8sSATokenValidator, err := createLegacyValidators(ctx, cfg, appMetrics, &logger)
+	if err != nil {
+		return err
+	}
 
 	// Only initialize validators if the gRPC server is enabled (port != 0)
 	if cfg.Server.Port != 0 {
-		// Create GitHub OIDC validator if enabled
-		if cfg.GitHubOIDC.Enabled {
-			v, err := githuboidc.NewValidator(ctx, cfg.GitHubOIDC, appMetrics, &logger)
-			if err != nil {
-				logger.Error("failed to create GitHub OIDC validator", zap.Error(err))
-				return err
-			}
-			// In-memory replay cache only protects against replay within this process. Multi-replica
-			// deployments need a shared backend (e.g. Redis) — a workload could otherwise replay the
-			// same token against a different replica. For now operators running >1 replica must
-			// serialize through a single instance, gate load-balancing to sticky routing, or accept
-			// the risk.
-			//
-			// TODO(replay-cache-backend): implement a pluggable ReplayCache backend (Redis first).
-			//   - Add a `replayCache` block to SpireIdentityExchangeConfig (kind: memory|redis, addr,
-			//     password, db, key prefix, ttl).
-			//   - Implement RedisReplayCache satisfying the existing cache.ReplayCache interface.
-			//   - Wire selection here based on cfg.ReplayCache.Kind; default remains in-memory.
-			//   - Add integration tests with miniredis. Cross-replica replay rejection must be
-			//     verified end-to-end (two SIE instances + one shared Redis).
-			//   - Drop the WARN log below once a shared backend is configured.
-			githubOIDCValidator = cache.NewReplayCheckingValidator(v, cache.NewInMemoryReplayCache(ctx))
-			logger.Info("GitHub OIDC validator enabled with in-memory replay cache")
-			logger.Warn("replay cache is in-memory only: multi-replica deployments can be bypassed by replaying a token against a different replica. Run a single replica until a shared backend is configured.")
-		}
-
-		// Create K8s SA token validator if enabled
-		if cfg.K8sSAToken.Enabled {
-			v, err := k8ssatoken.NewValidator(cfg.K8sSAToken, &logger)
-			if err != nil {
-				logger.Error("failed to create K8s SA token validator", zap.Error(err))
-				return err
-			}
-			k8sSATokenValidator = v
-			logger.Info("Kubernetes SA token validator enabled")
-		}
-
-		if githubOIDCValidator == nil && k8sSATokenValidator == nil && len(cfg.Auth.Plugins) == 0{
+		if githubOIDCValidator == nil && k8sSATokenValidator == nil && len(cfg.Auth.Plugins) == 0 {
 			logger.Error("at least one authentication method must be enabled")
 			return fmt.Errorf("no authentication method enabled")
 		}
-
 	} else {
 		logger.Info("gRPC port is 0; skipping token validator initializations")
 	}
@@ -160,7 +105,6 @@ func run() error {
 	// TODO: replay-cache wrap the pkg/validator instances too. The existing
 	// internal/cache.NewReplayCheckingValidator wraps the internal validator
 	// interface; equivalent wrapping for the pkg validator is a follow-up.
-
 	return service.Run(ctx, cfg, spireClient, githubOIDCValidator, k8sSATokenValidator, appMetrics, &logger)
 }
