+++ internal/service/runner_test.go
@@ -1,3 +1,5 @@
+//go:build legacy
+
 package service
 
 import (
